+++ src/main/java/com/supermartijn642/simplemagnets/SimpleMagnetsClient.java
@@ -9,22 +9,21 @@ import com.supermartijn642.simplemagnets.gui.MagnetContainerScreen;
 import com.supermartijn642.simplemagnets.packets.magnet.PacketToggleMagnet;
 import net.minecraft.client.KeyMapping;
 import net.minecraft.client.Minecraft;
-import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.client.event.InputEvent;
 import net.minecraftforge.client.event.RegisterKeyMappingsEvent;
-import net.minecraftforge.common.MinecraftForge;
-import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 
 /**
  * Created 7/7/2020 by SuperMartijn642
  */
-@Mod.EventBusSubscriber(bus = Mod.EventBusSubscriber.Bus.MOD, value = Dist.CLIENT)
 public class SimpleMagnetsClient {
 
     private static KeyMapping MAGNET_TOGGLE_KEY;
 
-    public static void register(){
+    public static void register(FMLJavaModLoadingContext context){
+        RegisterKeyMappingsEvent.getBus(context.getModBusGroup()).addListener(SimpleMagnetsClient::registerKeyBindings);
+        InputEvent.Key.BUS.addListener(SimpleMagnetsClient::onKey);
+
         ClientRegistrationHandler handler = ClientRegistrationHandler.get("simplemagnets");
         // Screens
         handler.registerContainerScreen(() -> SimpleMagnets.magnet_container, container -> WidgetContainerScreen.of(new MagnetContainerScreen(), container, false));
@@ -35,11 +34,9 @@ public class SimpleMagnetsClient {
         handler.registerCustomBlockEntityRenderer(() -> SimpleMagnets.advanced_demagnetization_coil_tile, DemagnetizationCoilBlockEntityRenderer::new);
     }
 
-    @SubscribeEvent
     public static void registerKeyBindings(RegisterKeyMappingsEvent e){
         MAGNET_TOGGLE_KEY = new KeyMapping("simplemagnets.keys.toggle", 72/*'h'*/, "simplemagnets.keys.category");
         e.register(MAGNET_TOGGLE_KEY);
-        MinecraftForge.EVENT_BUS.addListener(SimpleMagnetsClient::onKey);
     }
 
     public static void onKey(InputEvent.Key e){