+++ src/main/java/com/supermartijn642/simplemagnets/gui/CheckBox.java
@@ -1,7 +1,7 @@
 package com.supermartijn642.simplemagnets.gui;
 
 import com.supermartijn642.core.TextComponents;
-import com.supermartijn642.core.gui.ScreenUtils;
+import com.supermartijn642.core.gui.GuiGraphicsHelper;
 import com.supermartijn642.core.gui.widget.WidgetRenderContext;
 import com.supermartijn642.core.gui.widget.premade.AbstractButtonWidget;
 import com.supermartijn642.core.util.Holder;
@@ -16,7 +16,7 @@ import java.util.function.Function;
  */
 public class CheckBox extends AbstractButtonWidget {
 
-    private static final ResourceLocation BUTTONS = ResourceLocation.fromNamespaceAndPath("simplemagnets", "textures/checkmarkbox.png");
+    public static final ResourceLocation BUTTONS = ResourceLocation.fromNamespaceAndPath("simplemagnets", "checkmarkbox");
 
     private final Function<Boolean,String> translationKey;
     public boolean checked;
@@ -32,8 +32,8 @@ public class CheckBox extends AbstractButtonWidget {
     }
 
     @Override
-    public void render(WidgetRenderContext context, int mouseX, int mouseY){
-        ScreenUtils.drawTexture(BUTTONS, context.poseStack(), this.x, this.y - 3, this.width + 3, this.height + 3, this.checked ? 0 : 0.5f, (this.active ? this.isFocused() ? 1 : 0 : 2) / 3f, 0.5f, 1 / 3f);
+    public void render(WidgetRenderContext context, GuiGraphicsHelper graphics, int mouseX, int mouseY){
+        graphics.submitSprite(BUTTONS, this.x, this.y - 3, this.width + 3, this.height + 3, p -> p.uv(this.checked ? 0 : 0.5f, (this.active ? this.isFocused() ? 1 : 0 : 2) / 3f, 0.5f, 1 / 3f));
     }
 
     @Override